reorganize features to use new react hooks
