+++ src/features/combat/reducer.js
@@ -7,7 +7,6 @@ const combatReducer = (state = initialState, action) => {
     case "START_COMBAT":
       return {
         visible: true,
-        myHP: 100,
         ...action.payload
       };
     case "END_COMBAT":
